ai-apis-and-standalone-apps/assignments/Assignment12/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0